Added plugin_data
Tabs are now split between "Plugin" and "Admin" types.  Settings types
are kept inside the form which updates the plugin options in the
database.  General tabs are for forms and other fields we don't want to
store directly in the site meta but rather in custom tables we create
for our plugin - these forms use javascript defined in the admin.js file
to post their json to be processed.