Added pretty filename and regex replace
Pretty filename strips some of the standard file characters like _’s
and -’s and the file extension and replaces them with a space.  This is
activated by setting the prettyfilename parameter equal to anything but
false.  The regexstrip allows the user to pipe in some regex to strip
from the filename - this will throw an error if the regex isn’t
formatted properly…